// apps/cli/src/types/Birdeye.ts
export const WalletAnalyticsIntervalSchema = z.enum(['1h', '1d'])
// ...
export const WalletAnalyticsNetWorthCommandOptionsSchema = z.object({
  wallet: z.string().min(1),
  limit: z.number().int().min(1).max(100).nullish(),
  out: z.string().nullish()
})
export type WalletAnalyticsNetWorthCommandOptions = z.infer<
  typeof WalletAnalyticsNetWorthCommandOptionsSchema
>

export const WalletAnalyticsNetWorthDetailsCommandOptionsSchema = z.object({
  wallet: z.string().min(1),
  type: WalletAnalyticsIntervalSchema.nullish(),
  time: z.string().min(1).nullish(),
  limit: z.number().int().min(1).max(100).nullish(),
  out: z.string().nullish()
})
export type WalletAnalyticsNetWorthDetailsCommandOptions = z.infer<
  typeof WalletAnalyticsNetWorthDetailsCommandOptionsSchema
>

export const WalletAnalyticsNetWorthChartCommandOptionsSchema = z.object({
  wallet: z.string().min(1),
  type: WalletAnalyticsIntervalSchema.nullish(),
  count: z.number().int().min(1).max(30).nullish(),
  out: z.string().nullish()
})
export type WalletAnalyticsNetWorthChartCommandOptions = z.infer<
  typeof WalletAnalyticsNetWorthChartCommandOptionsSchema
>

export const WalletAnalyticsBalanceChangeCommandOptionsSchema = z.object({
  wallet: z.string().min(1),
  from: z.number().int().positive().nullish(),
  to: z.number().int().positive().nullish(),
  limit: z.number().int().min(1).max(100).nullish(),
  out: z.string().nullish()
})
export type WalletAnalyticsBalanceChangeCommandOptions = z.infer<
  typeof WalletAnalyticsBalanceChangeCommandOptionsSchema
>

export const WalletAnalyticsTransferTotalCommandOptionsSchema = z.object({
  wallet: z.string().min(1),
  out: z.string().nullish()
})